requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
dataclasses==0.6
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.7
django-unixdatetimefield==1.0.3
djangorestframework==3.9.2
djongo==1.2.32
get==2018.11.19
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
lxml==4.3.2
oauthlib==3.0.1
parsel==1.5.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
post==2018.11.20
prompt-toolkit==2.0.9
ptyprocess==0.6.0
public==2018.11.20
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==3.7.2
pyOpenSSL==19.0.0
python3-openid==3.1.0
pytz==2018.9
query-string==2018.11.20
queuelib==1.5.0
request==2018.11.20
requests==2.21.0
requests-oauthlib==1.2.0
Scrapy==1.6.0
service-identity==18.1.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.2.4
traitlets==4.3.2
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.20.0
wcwidth==0.1.7
zope.interface==4.6.0
